requirements.txt
streamlit==1.22.0
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0
numpy==1.24.2
opencv-python==4.7.0.72
matplotlib==3.7.0
lime==0.2.0.1
captum==0.6.0
scikit-image==0.19.3